requirements.txt
# Database and web framework requirements
Flask==3.1.2
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
python-engineio==4.13.0
python-socketio==5.16.0

# LLMs requirements
# Semantic Kernel for orchestration with function calling support
semantic-kernel>=1.18.0
# OpenAI API (for main orchestration and Generator Agent)
openai>=1.0.0
# Anthropic Claude API (for Detector Agent)
anthropic>=0.18.0
# Google Gemini API (for Judge Agent)
google.genai==1.39.0
# Environment variables
python-dotenv>=1.2.1
# Async support (included with other packages, but explicit for clarity)
aiohttp>=3.9.0
# Calculate LLM cost
tokencost>=0.1.26